MovementSpeed changed to float

+++ Assets/Gameplay/Scripts/UnitStats.cs
@@ -5,14 +5,14 @@ using System.Collections;
 public class UnitStats {
 
     [SerializeField] int damage;
-    [SerializeField] int movementSpeed;
+	[SerializeField] float movementSpeed;
     private int currentHealth;
     [SerializeField] int maxHealth;
 	[SerializeField] int experience;
 	[SerializeField] int itemDropChance;
 
     int bonusDamage;
-    int bonusMovementSpeed;
+	float bonusMovementSpeed;
     int bonusMaxHealth;
 	int bonusItemDropChance;
 
@@ -23,14 +23,14 @@ public class UnitStats {
 	int bonusIncreasedMagazineSize; // %
 
 	int statPointsMaxHealth;
-	int statPointsMovementSpeed;
+	float statPointsMovementSpeed;
 	int statPointsItemDropChance;
 	float statPointsReducedReloadTime;
 	float statPointsIncreasedBonusDuration;
 
 
     public int Damage{ get {return damage + bonusDamage;}}
-	public int MovementSpeed{ get {return movementSpeed + bonusMovementSpeed + statPointsMovementSpeed;}}
+	public float MovementSpeed{ get {return movementSpeed + bonusMovementSpeed + statPointsMovementSpeed;}}
     public int CurrentHealth{ get {return currentHealth;} set{currentHealth = value;}}
 	public int MaxHealth{ get {return maxHealth + bonusMaxHealth + statPointsMaxHealth;}}
 	public int Experience{ get {return experience;} set{ experience = value;}}
@@ -41,12 +41,12 @@ public class UnitStats {
 	public float IncreasedMagazineSize{get {return bonusIncreasedMagazineSize;}}
 
     public int BaseDamage{ get {return damage;} set{damage = value;}}
-    public int BaseMovementSpeed{ get {return movementSpeed;} set{movementSpeed = value;}}
+    public float BaseMovementSpeed{ get {return movementSpeed;} set{movementSpeed = value;}}
     public int BaseMaxHealth{ get {return maxHealth;} set{maxHealth = value;}}
     public int BaseItemDropChance{ get {return itemDropChance;} set{itemDropChance = value;}}
 
     public int BonusDamage{ get {return bonusDamage;} set{bonusDamage = value;}}
-    public int BonusMovementSpeed{ get {return bonusMovementSpeed;} set{bonusMovementSpeed = value;}}
+    public float BonusMovementSpeed{ get {return bonusMovementSpeed;} set{bonusMovementSpeed = value;}}
     public int BonusMaxHealth{ get {return bonusMaxHealth;} set{bonusMaxHealth = value;}}
 	public int BonusExperienceGained{ get {return bonusExperienceGained;} set{bonusExperienceGained = value;}}
     public int BonusItemDropChance{ get {return bonusItemDropChance;} set{bonusItemDropChance = value;}}
@@ -55,7 +55,7 @@ public class UnitStats {
 	public int BonusIncreasedMagazineSize{get {return bonusIncreasedMagazineSize;} set{ bonusIncreasedMagazineSize = value;}}
 
 	public int StatPointsMaxHealth{ get {return statPointsMaxHealth;} set{statPointsMaxHealth = value;}}
-	public int StatPointsMovementSpeed{ get {return statPointsMovementSpeed;} set{statPointsMovementSpeed = value;}}
+	public float StatPointsMovementSpeed{ get {return statPointsMovementSpeed;} set{statPointsMovementSpeed = value;}}
 	public int StatPointsItemDropChance{ get {return statPointsItemDropChance;} set{statPointsItemDropChance = value;}}
 	public float StatPointsReducedReloadTime{ get {return statPointsReducedReloadTime;} set{statPointsReducedReloadTime = value;}}
 	public float StatPointsIncreasedBonusDuration{ get {return statPointsIncreasedBonusDuration;} set{statPointsIncreasedBonusDuration = value;}}
